requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
numpy>=1.25.0
plotly>=5.20.0
requests>=2.31.0
pytrends>=4.9.2
python-dotenv>=1.0.1
